@@ -10,30 +10,27 @@
     "watch": "gulp watch",
     "serve": "gulp serve"
   },
-  "author": "Eisneim <eisneim1@gmail.com>",
+  "author": "Andy <minzheng1988@aliyun.com>",
   "license": "MIT",
   "devDependencies": {
-    "gulp": "^3.8.11",
-    "gulp-sass": "^1.2.4",
-    "gulp-uglify": "~0.2.1",
-    "gulp-util": "~2.2.14",
-    "gulp-minify-css": "^0.3.4",
-    "gulp-autoprefixer": "0.0.7",
+    "babelify": "^6.3.0",
+    "gulp": "^3.9.0",
+    "gulp-autoprefixer": "^3.0.1",
     "gulp-browserify": "^0.5.1",
-    "gulp-concat": "^2.5.2",
-    "gulp-livereload": "^2.1.0",
-    "gulp-open": "^0.3.2",
-    "gulp-react": "^2.0.0",
-    "babel-runtime":"^5.1.11",
-    "babelify":"^6.0.2"
+    "gulp-concat": "^2.6.0",
+    "gulp-connect": "^2.2.0",
+    "gulp-livereload": "^3.8.0",
+    "gulp-minify-css": "^1.2.1",
+    "gulp-sass": "^2.0.4",
+    "gulp-uglify": "^1.4.0"
   },
   "dependencies": {
+    "classnames": "^2.1.3",
     "dragdealer": "^0.9.8",
-    "flux": "^2.0.0",
     "lodash": "^3.3.1",
-    "react": "^0.12.2",
-    "react-document-title": "^1.0.2",
-    "react-router": "^0.12.4",
-    "superagent": "^1.1.0"
+    "object-assign": "^4.0.1",
+    "react": "^0.13.3",
+    "flux": "^2.1.1",
+    "react-router": "^0.13.3"
   }
 }
